Move ReST API controller to the controller package
